versions: add build_revision and enable by default
Add `sw/share/ogygia/build-revision` to the system closure including the
hash the closure was built from.

This makes the closure different every time the Git hash changes, but is
incredibly useful. This builds the foundations of my existing version
tracking system and allows you to track the current, booted, and
nextboot versions. Assuming all closures are built without `--impure`
and good hygiene is used in other places this allows you to track
exactly what every machine in the island is running.

Adding this tracking ability to all islands by default is a precursor to
more comprehensive island tracking in the future.

Test plan:
- Use in JakeHillion/nixos and switch version tracking tooling to it[0].
  All version tracking tooling continues to work.

[0] https://gitea.hillion.co.uk/JakeHillion/nixos/pulls/739